+++ src/main/java/insane96mcp/shieldsplus/world/item/enchantment/AblazeEnchantment.java
@@ -1,13 +1,10 @@
 package insane96mcp.shieldsplus.world.item.enchantment;
 
-import insane96mcp.insanelib.world.enchantments.IEnchantmentTooltip;
 import insane96mcp.insanelib.world.scheduled.ScheduledTasks;
 import insane96mcp.insanelib.world.scheduled.ScheduledTickTask;
 import insane96mcp.shieldsplus.module.BaseFeature;
 import insane96mcp.shieldsplus.setup.SPEnchantments;
 import insane96mcp.shieldsplus.world.item.SPShieldItem;
-import net.minecraft.ChatFormatting;
-import net.minecraft.network.chat.Component;
 import net.minecraft.world.damagesource.DamageSource;
 import net.minecraft.world.entity.Entity;
 import net.minecraft.world.entity.EquipmentSlot;
@@ -17,7 +14,7 @@ import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.enchantment.Enchantment;
 import net.minecraftforge.event.entity.living.ShieldBlockEvent;
 
-public class AblazeEnchantment extends Enchantment implements IBlockingEffect, IEnchantmentTooltip {
+public class AblazeEnchantment extends Enchantment implements IBlockingEffect {
 
 	public AblazeEnchantment() {
 		super(Rarity.UNCOMMON, SPShieldItem.ENCHANTMENT_CATEGORY, new EquipmentSlot[]{EquipmentSlot.MAINHAND, EquipmentSlot.OFFHAND});
@@ -55,10 +52,5 @@ public class AblazeEnchantment extends Enchantment implements IBlockingEffect, I
 					}
 				}
 			});
-	}
-
-	@Override
-	public Component getTooltip(ItemStack itemStack, int lvl) {
-		return Component.translatable(this.getDescriptionId() + ".tooltip", BaseFeature.enchantmentsAblazeTimeOnFire * lvl).withStyle(ChatFormatting.DARK_PURPLE);
 	}
 }