+++ src/main/java/insane96mcp/shieldsplus/world/item/enchantment/AegisEnchantment.java
@@ -1,19 +1,15 @@
 package insane96mcp.shieldsplus.world.item.enchantment;
 
-import insane96mcp.insanelib.InsaneLib;
-import insane96mcp.insanelib.world.enchantments.IEnchantmentTooltip;
 import insane96mcp.shieldsplus.module.BaseFeature;
 import insane96mcp.shieldsplus.setup.SPEnchantments;
 import insane96mcp.shieldsplus.world.item.SPShieldItem;
-import net.minecraft.ChatFormatting;
-import net.minecraft.network.chat.Component;
 import net.minecraft.world.entity.EquipmentSlot;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.enchantment.Enchantment;
 import net.minecraftforge.event.entity.living.LivingDamageEvent;
 import org.jetbrains.annotations.NotNull;
 
-public class AegisEnchantment extends Enchantment implements IEnchantmentTooltip {
+public class AegisEnchantment extends Enchantment {
 	public AegisEnchantment() {
 		super(Rarity.COMMON, SPShieldItem.ENCHANTMENT_CATEGORY, new EquipmentSlot[]{EquipmentSlot.MAINHAND, EquipmentSlot.OFFHAND});
 	}
@@ -74,10 +70,5 @@ public class AegisEnchantment extends Enchantment implements IEnchantmentTooltip
 			return;
 
 		event.setAmount(event.getAmount() * (1f - AegisEnchantment.getResistance(lvl)));
-	}
-
-	@Override
-	public Component getTooltip(ItemStack itemStack, int lvl) {
-		return Component.translatable(this.getDescriptionId() + ".tooltip", InsaneLib.ONE_DECIMAL_FORMATTER.format(getResistance(lvl) * 100f)).withStyle(ChatFormatting.DARK_PURPLE);
 	}
 }