+++ src/main/java/insane96mcp/shieldsplus/world/item/enchantment/FastRecoveryEnchantment.java
@@ -1,16 +1,11 @@
 package insane96mcp.shieldsplus.world.item.enchantment;
 
-import insane96mcp.insanelib.InsaneLib;
-import insane96mcp.insanelib.world.enchantments.IEnchantmentTooltip;
 import insane96mcp.shieldsplus.module.BaseFeature;
 import insane96mcp.shieldsplus.world.item.SPShieldItem;
-import net.minecraft.ChatFormatting;
-import net.minecraft.network.chat.Component;
 import net.minecraft.world.entity.EquipmentSlot;
-import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.enchantment.Enchantment;
 
-public class FastRecoveryEnchantment extends Enchantment implements IEnchantmentTooltip {
+public class FastRecoveryEnchantment extends Enchantment {
 	public FastRecoveryEnchantment() {
 		super(Rarity.RARE, SPShieldItem.ENCHANTMENT_CATEGORY, new EquipmentSlot[]{EquipmentSlot.MAINHAND, EquipmentSlot.OFFHAND});
 	}
@@ -25,10 +20,5 @@ public class FastRecoveryEnchantment extends Enchantment implements IEnchantment
 
 	public static float getCooldownReduction(int lvl) {
 		return BaseFeature.enchantmentsFastRecoveryCooldownReduction.floatValue();
-	}
-
-	@Override
-	public Component getTooltip(ItemStack itemStack, int lvl) {
-		return Component.translatable(this.getDescriptionId() + ".tooltip", InsaneLib.ONE_DECIMAL_FORMATTER.format(getCooldownReduction(lvl) * 100f)).withStyle(ChatFormatting.DARK_PURPLE);
 	}
 }