+++ src/main/java/insane96mcp/shieldsplus/world/item/enchantment/ReflectionEnchantment.java
@@ -1,20 +1,15 @@
 package insane96mcp.shieldsplus.world.item.enchantment;
 
-import insane96mcp.insanelib.InsaneLib;
-import insane96mcp.insanelib.world.enchantments.IEnchantmentTooltip;
 import insane96mcp.shieldsplus.module.BaseFeature;
 import insane96mcp.shieldsplus.world.item.SPShieldItem;
-import net.minecraft.ChatFormatting;
-import net.minecraft.network.chat.Component;
 import net.minecraft.world.damagesource.DamageSource;
 import net.minecraft.world.entity.EquipmentSlot;
 import net.minecraft.world.entity.LivingEntity;
-import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.enchantment.Enchantment;
 import net.minecraftforge.event.entity.living.ShieldBlockEvent;
 import org.jetbrains.annotations.NotNull;
 
-public class ReflectionEnchantment extends Enchantment implements IBlockingEffect, IEnchantmentTooltip {
+public class ReflectionEnchantment extends Enchantment implements IBlockingEffect {
 	public ReflectionEnchantment() {
 		super(Rarity.RARE, SPShieldItem.ENCHANTMENT_CATEGORY, new EquipmentSlot[]{EquipmentSlot.MAINHAND, EquipmentSlot.OFFHAND});
 	}
@@ -46,10 +41,5 @@ public class ReflectionEnchantment extends Enchantment implements IBlockingEffec
 			return;
 
 		sourceEntity.hurt(blockingEntity.damageSources().thorns(blockingEntity), ReflectionEnchantment.getReflectedDamage(lvl) * amount);
-	}
-
-	@Override
-	public Component getTooltip(ItemStack itemStack, int lvl) {
-		return Component.translatable(this.getDescriptionId() + ".tooltip", InsaneLib.ONE_DECIMAL_FORMATTER.format(getReflectedDamage(lvl) * 100f)).withStyle(ChatFormatting.DARK_PURPLE);
 	}
 }