+++ changelog.md
@@ -2,6 +2,10 @@
 
 ## Upcoming
 * Added Celestial Guardian treasure enchantment
+* Perfect parry
+  * The attacking entity (if melee) now gets "stunned"
+  * Reduced window to activate to 1 tick (from 2)
+* Added support for IguanaTweaks Reborn enchantments descriptions
 
 ## 1.9.0
 * Added Perfect Parry enchantment